Settings moved to self file
